Split up ForkChoiceUtil and move away from static methods
