+++ ethereum/core/src/main/java/tech/pegasys/teku/core/ForkChoiceUtil.java
@@ -13,37 +13,20 @@
 
 package tech.pegasys.teku.core;
 
-import static com.google.common.base.Preconditions.checkArgument;
-import static tech.pegasys.teku.datastructures.util.AttestationProcessingResult.SUCCESSFUL;
-import static tech.pegasys.teku.datastructures.util.AttestationUtil.is_valid_indexed_attestation;
 import static tech.pegasys.teku.datastructures.util.BeaconStateUtil.compute_epoch_at_slot;
 import static tech.pegasys.teku.datastructures.util.BeaconStateUtil.compute_start_slot_at_epoch;
-import static tech.pegasys.teku.datastructures.util.BeaconStateUtil.get_block_root_at_slot;
-import static tech.pegasys.teku.util.config.Constants.GENESIS_EPOCH;
 import static tech.pegasys.teku.util.config.Constants.GENESIS_SLOT;
-import static tech.pegasys.teku.util.config.Constants.SAFE_SLOTS_TO_UPDATE_JUSTIFIED;
 import static tech.pegasys.teku.util.config.Constants.SECONDS_PER_SLOT;
 
 import java.time.Instant;
 import java.util.NavigableMap;
 import java.util.Optional;
 import java.util.TreeMap;
-import javax.annotation.CheckReturnValue;
 import org.apache.tuweni.bytes.Bytes32;
-import tech.pegasys.teku.core.lookup.IndexedAttestationProvider;
-import tech.pegasys.teku.core.results.BlockImportResult;
-import tech.pegasys.teku.data.BlockProcessingRecord;
-import tech.pegasys.teku.datastructures.attestation.ValidateableAttestation;
-import tech.pegasys.teku.datastructures.blocks.BeaconBlock;
-import tech.pegasys.teku.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.datastructures.forkchoice.MutableStore;
+import tech.pegasys.teku.datastructures.forkchoice.ReadOnlyForkChoiceStrategy;
 import tech.pegasys.teku.datastructures.forkchoice.ReadOnlyStore;
-import tech.pegasys.teku.datastructures.operations.Attestation;
-import tech.pegasys.teku.datastructures.state.BeaconState;
-import tech.pegasys.teku.datastructures.state.Checkpoint;
-import tech.pegasys.teku.datastructures.util.AttestationProcessingResult;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.protoarray.ForkChoiceStrategy;
 
 public class ForkChoiceUtil {
 
@@ -86,12 +69,12 @@ public class ForkChoiceUtil {
    *     <a>https://github.com/ethereum/eth2.0-specs/blob/v0.10.1/specs/phase0/fork-choice.md#get_ancestor</a>
    */
   public static Optional<Bytes32> get_ancestor(
-      ForkChoiceStrategy forkChoiceStrategy, Bytes32 root, UInt64 slot) {
+      ReadOnlyForkChoiceStrategy forkChoiceStrategy, Bytes32 root, UInt64 slot) {
     return forkChoiceStrategy.getAncestor(root, slot);
   }
 
   public static NavigableMap<UInt64, Bytes32> getAncestors(
-      ForkChoiceStrategy forkChoiceStrategy,
+      ReadOnlyForkChoiceStrategy forkChoiceStrategy,
       Bytes32 root,
       UInt64 startSlot,
       UInt64 step,
@@ -118,7 +101,7 @@ public class ForkChoiceUtil {
    *     backwards
    */
   public static NavigableMap<UInt64, Bytes32> getAncestorsOnFork(
-      ForkChoiceStrategy forkChoiceStrategy, Bytes32 root, UInt64 startSlot) {
+      ReadOnlyForkChoiceStrategy forkChoiceStrategy, Bytes32 root, UInt64 startSlot) {
     final NavigableMap<UInt64, Bytes32> roots = new TreeMap<>();
     Bytes32 parentRoot = root;
     Optional<UInt64> parentSlot = forkChoiceStrategy.blockSlot(parentRoot);
@@ -145,31 +128,6 @@ public class ForkChoiceUtil {
             roots.put(slot, root);
           }
         });
-  }
-
-  /*
-  To address the bouncing attack, only update conflicting justified
-  checkpoints in the fork choice if in the early slots of the epoch.
-  Otherwise, delay incorporation of new justified checkpoint until next epoch boundary.
-  See https://ethresear.ch/t/prevention-of-bouncing-attack-on-ffg/6114 for more detailed analysis and discussion.
-  */
-
-  private static boolean should_update_justified_checkpoint(
-      ReadOnlyStore store,
-      Checkpoint new_justified_checkpoint,
-      ForkChoiceStrategy forkChoiceStrategy) {
-    if (compute_slots_since_epoch_start(get_current_slot(store, true))
-            .compareTo(UInt64.valueOf(SAFE_SLOTS_TO_UPDATE_JUSTIFIED))
-        < 0) {
-      return true;
-    }
-
-    UInt64 justifiedSlot = compute_start_slot_at_epoch(store.getJustifiedCheckpoint().getEpoch());
-    return hasAncestorAtSlot(
-        forkChoiceStrategy,
-        new_justified_checkpoint.getRoot(),
-        justifiedSlot,
-        store.getJustifiedCheckpoint().getRoot());
   }
 
   // Fork Choice Event Handlers
@@ -207,248 +165,5 @@ public class ForkChoiceUtil {
         > 0) {
       store.setJustifiedCheckpoint(store.getBestJustifiedCheckpoint());
     }
-  }
-
-  /**
-   * Perform block processing. The supplied blockSlotState must already have empty slots processed
-   * to the same slot as the block.
-   */
-  @CheckReturnValue
-  public static BlockImportResult on_block(
-      final MutableStore store,
-      final SignedBeaconBlock signed_block,
-      BeaconState blockSlotState,
-      final StateTransition st,
-      final ForkChoiceStrategy forkChoiceStrategy,
-      final IndexedAttestationProvider indexedAttestationProvider) {
-    checkArgument(
-        blockSlotState.getSlot().equals(signed_block.getSlot()),
-        "State must have slots processed up to the block slot");
-    final BeaconBlock block = signed_block.getMessage();
-
-    // Return early if precondition checks fail;
-    final Optional<BlockImportResult> maybeFailure =
-        checkOnBlockConditions(block, blockSlotState, store, forkChoiceStrategy);
-    if (maybeFailure.isPresent()) {
-      return maybeFailure.get();
-    }
-
-    // Make a copy of the state to avoid mutability issues
-    BeaconState state;
-
-    // Check the block is valid and compute the post-state
-    try {
-      state =
-          st.processAndValidateBlock(
-              signed_block, indexedAttestationProvider, blockSlotState, true);
-    } catch (StateTransitionException e) {
-      return BlockImportResult.failedStateTransition(e);
-    }
-
-    // Add new block to store
-    store.putBlockAndState(signed_block, state);
-
-    // Update justified checkpoint
-    final Checkpoint justifiedCheckpoint = state.getCurrent_justified_checkpoint();
-    if (justifiedCheckpoint.getEpoch().compareTo(store.getJustifiedCheckpoint().getEpoch()) > 0) {
-      if (justifiedCheckpoint.getEpoch().compareTo(store.getBestJustifiedCheckpoint().getEpoch())
-          > 0) {
-        store.setBestJustifiedCheckpoint(justifiedCheckpoint);
-      }
-      if (should_update_justified_checkpoint(store, justifiedCheckpoint, forkChoiceStrategy)) {
-        store.setJustifiedCheckpoint(justifiedCheckpoint);
-      }
-    }
-
-    // Update finalized checkpoint
-    final Checkpoint finalizedCheckpoint = state.getFinalized_checkpoint();
-    if (finalizedCheckpoint.getEpoch().compareTo(store.getFinalizedCheckpoint().getEpoch()) > 0) {
-      store.setFinalizedCheckpoint(finalizedCheckpoint);
-
-      // Potentially update justified if different from store
-      if (!store.getJustifiedCheckpoint().equals(state.getCurrent_justified_checkpoint())) {
-        // Update justified if new justified is later than store justified
-        // or if store justified is not in chain with finalized checkpoint
-        if (state
-                    .getCurrent_justified_checkpoint()
-                    .getEpoch()
-                    .compareTo(store.getJustifiedCheckpoint().getEpoch())
-                > 0
-            || !isFinalizedAncestorOfJustified(forkChoiceStrategy, store)) {
-          store.setJustifiedCheckpoint(state.getCurrent_justified_checkpoint());
-        }
-      }
-    }
-
-    final BlockProcessingRecord record = new BlockProcessingRecord(signed_block, state);
-    return BlockImportResult.successful(record);
-  }
-
-  private static boolean isFinalizedAncestorOfJustified(
-      ForkChoiceStrategy forkChoiceStrategy, ReadOnlyStore store) {
-    UInt64 finalizedSlot = store.getFinalizedCheckpoint().getEpochStartSlot();
-    return hasAncestorAtSlot(
-        forkChoiceStrategy,
-        store.getJustifiedCheckpoint().getRoot(),
-        finalizedSlot,
-        store.getFinalizedCheckpoint().getRoot());
-  }
-
-  private static Optional<BlockImportResult> checkOnBlockConditions(
-      final BeaconBlock block,
-      final BeaconState blockSlotState,
-      final ReadOnlyStore store,
-      final ForkChoiceStrategy forkChoiceStrategy) {
-    final UInt64 blockSlot = block.getSlot();
-    if (blockSlotState == null) {
-      return Optional.of(BlockImportResult.FAILED_UNKNOWN_PARENT);
-    }
-    if (!blockSlotState.getSlot().equals(blockSlot)) {
-      return Optional.of(BlockImportResult.FAILED_INVALID_ANCESTRY);
-    }
-    if (blockSlot.isGreaterThan(UInt64.valueOf(GENESIS_SLOT))
-        && !get_block_root_at_slot(blockSlotState, blockSlot.minus(1))
-            .equals(block.getParentRoot())) {
-      // Block is at same slot as its parent or the parent root doesn't match the state
-      return Optional.of(BlockImportResult.FAILED_INVALID_ANCESTRY);
-    }
-    if (blockIsFromFuture(store, blockSlot)) {
-      return Optional.of(BlockImportResult.FAILED_BLOCK_IS_FROM_FUTURE);
-    }
-    if (!blockDescendsFromLatestFinalizedBlock(block, store, forkChoiceStrategy)) {
-      return Optional.of(BlockImportResult.FAILED_INVALID_ANCESTRY);
-    }
-    return Optional.empty();
-  }
-
-  private static boolean blockIsFromFuture(ReadOnlyStore store, final UInt64 blockSlot) {
-    return get_current_slot(store).compareTo(blockSlot) < 0;
-  }
-
-  private static boolean hasAncestorAtSlot(
-      ForkChoiceStrategy forkChoiceStrategy, Bytes32 root, UInt64 slot, Bytes32 ancestorRoot) {
-    return get_ancestor(forkChoiceStrategy, root, slot)
-        .map(ancestorAtSlot -> ancestorAtSlot.equals(ancestorRoot))
-        .orElse(false);
-  }
-
-  public static boolean blockDescendsFromLatestFinalizedBlock(
-      final BeaconBlock block,
-      final ReadOnlyStore store,
-      final ForkChoiceStrategy forkChoiceStrategy) {
-    final Checkpoint finalizedCheckpoint = store.getFinalizedCheckpoint();
-    final UInt64 blockSlot = block.getSlot();
-
-    // Make sure this block's slot is after the latest finalized slot
-    return blockIsAfterLatestFinalizedSlot(blockSlot, finalizedCheckpoint.getEpochStartSlot())
-        && hasAncestorAtSlot(
-            forkChoiceStrategy,
-            block.getParentRoot(),
-            finalizedCheckpoint.getEpochStartSlot(),
-            finalizedCheckpoint.getRoot());
-  }
-
-  private static boolean blockIsAfterLatestFinalizedSlot(
-      final UInt64 blockSlot, final UInt64 finalizedEpochStartSlot) {
-    if (blockSlot.compareTo(finalizedEpochStartSlot) <= 0) {
-      return false;
-    } else {
-      return true;
-    }
-  }
-
-  @CheckReturnValue
-  public static AttestationProcessingResult validateAttestation(
-      final ReadOnlyStore store,
-      final ValidateableAttestation validateableAttestation,
-      final Optional<BeaconState> maybeTargetState,
-      final ForkChoiceStrategy forkChoiceStrategy) {
-    Attestation attestation = validateableAttestation.getAttestation();
-
-    return validateOnAttestation(store, attestation, forkChoiceStrategy)
-        .ifSuccessful(
-            () -> {
-              if (maybeTargetState.isEmpty()) {
-                return AttestationProcessingResult.UNKNOWN_BLOCK;
-              } else {
-                return is_valid_indexed_attestation(
-                    maybeTargetState.get(), validateableAttestation);
-              }
-            })
-        .ifSuccessful(() -> checkIfAttestationShouldBeSavedForFuture(store, attestation));
-  }
-
-  private static AttestationProcessingResult validateOnAttestation(
-      final ReadOnlyStore store,
-      final Attestation attestation,
-      final ForkChoiceStrategy forkChoiceStrategy) {
-    final Checkpoint target = attestation.getData().getTarget();
-    UInt64 current_epoch = compute_epoch_at_slot(get_current_slot(store));
-
-    // Use GENESIS_EPOCH for previous when genesis to avoid underflow
-    UInt64 previous_epoch =
-        current_epoch.compareTo(UInt64.valueOf(GENESIS_EPOCH)) > 0
-            ? current_epoch.minus(UInt64.ONE)
-            : UInt64.valueOf(GENESIS_EPOCH);
-
-    if (!target.getEpoch().equals(previous_epoch) && !target.getEpoch().equals(current_epoch)) {
-      return AttestationProcessingResult.invalid(
-          "Attestations must be from the current or previous epoch");
-    }
-
-    if (!target.getEpoch().equals(compute_epoch_at_slot(attestation.getData().getSlot()))) {
-      return AttestationProcessingResult.invalid("Attestation slot must be within specified epoch");
-    }
-
-    if (!forkChoiceStrategy.contains(target.getRoot())) {
-      // Attestations target must be for a known block. If a target block is unknown, delay
-      // consideration until the block is found
-      return AttestationProcessingResult.UNKNOWN_BLOCK;
-    }
-
-    Optional<UInt64> blockSlot =
-        forkChoiceStrategy.blockSlot(attestation.getData().getBeacon_block_root());
-    if (blockSlot.isEmpty()) {
-      // Attestations must be for a known block. If block is unknown, delay consideration until the
-      // block is found
-      return AttestationProcessingResult.UNKNOWN_BLOCK;
-    }
-
-    if (blockSlot.get().compareTo(attestation.getData().getSlot()) > 0) {
-      return AttestationProcessingResult.invalid(
-          "Attestations must not be for blocks in the future. If not, the attestation should not be considered");
-    }
-
-    // LMD vote must be consistent with FFG vote target
-    final UInt64 target_slot = compute_start_slot_at_epoch(target.getEpoch());
-    if (get_ancestor(forkChoiceStrategy, attestation.getData().getBeacon_block_root(), target_slot)
-        .map(ancestorRoot -> !ancestorRoot.equals(target.getRoot()))
-        .orElse(true)) {
-      return AttestationProcessingResult.invalid(
-          "LMD vote must be consistent with FFG vote target");
-    }
-
-    return SUCCESSFUL;
-  }
-
-  private static AttestationProcessingResult checkIfAttestationShouldBeSavedForFuture(
-      ReadOnlyStore store, Attestation attestation) {
-
-    // Attestations can only affect the fork choice of subsequent slots.
-    // Delay consideration in the fork choice until their slot is in the past.
-    final UInt64 currentSlot = get_current_slot(store);
-    if (currentSlot.compareTo(attestation.getData().getSlot()) < 0) {
-      return AttestationProcessingResult.SAVED_FOR_FUTURE;
-    }
-    if (currentSlot.compareTo(attestation.getData().getSlot()) == 0) {
-      return AttestationProcessingResult.DEFER_FOR_FORK_CHOICE;
-    }
-
-    // Attestations cannot be from future epochs. If they are, delay consideration until the epoch
-    // arrives
-    if (currentSlot.compareTo(attestation.getData().getTarget().getEpochStartSlot()) < 0) {
-      return AttestationProcessingResult.SAVED_FOR_FUTURE;
-    }
-    return SUCCESSFUL;
   }
 }
