+++ protoarray/src/main/java/tech/pegasys/teku/protoarray/ForkChoiceStrategy.java
@@ -13,16 +13,14 @@
 
 package tech.pegasys.teku.protoarray;
 
-import java.util.Map;
-import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.datastructures.forkchoice.MutableStore;
+import tech.pegasys.teku.datastructures.forkchoice.ReadOnlyForkChoiceStrategy;
 import tech.pegasys.teku.datastructures.operations.IndexedAttestation;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.datastructures.state.Checkpoint;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public interface ForkChoiceStrategy {
+public interface ForkChoiceStrategy extends ReadOnlyForkChoiceStrategy {
 
   Bytes32 findHead(
       final MutableStore store,
@@ -31,14 +29,4 @@ public interface ForkChoiceStrategy {
       final BeaconState justifiedCheckpointState);
 
   void onAttestation(final MutableStore store, final IndexedAttestation attestation);
-
-  Optional<UInt64> blockSlot(Bytes32 blockRoot);
-
-  Optional<Bytes32> blockParentRoot(Bytes32 blockRoot);
-
-  Optional<Bytes32> getAncestor(Bytes32 blockRoot, UInt64 slot);
-
-  Map<Bytes32, UInt64> getChainHeads();
-
-  boolean contains(Bytes32 blockRoot);
 }
