+++ storage/src/main/java/tech/pegasys/teku/storage/client/RecentChainData.java
@@ -37,6 +37,7 @@ import tech.pegasys.teku.datastructures.blocks.BeaconBlock;
 import tech.pegasys.teku.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.datastructures.blocks.StateAndBlockSummary;
+import tech.pegasys.teku.datastructures.forkchoice.ReadOnlyForkChoiceStrategy;
 import tech.pegasys.teku.datastructures.genesis.GenesisData;
 import tech.pegasys.teku.datastructures.state.AnchorPoint;
 import tech.pegasys.teku.datastructures.state.BeaconState;
@@ -455,7 +456,7 @@ public abstract class RecentChainData implements StoreUpdateHandler {
 
   public Map<Bytes32, UInt64> getChainHeads() {
     return getForkChoiceStrategy()
-        .map(ForkChoiceStrategy::getChainHeads)
+        .map(ReadOnlyForkChoiceStrategy::getChainHeads)
         .orElse(Collections.emptyMap());
   }
 }
